+++ test/common.nix
@@ -174,7 +174,8 @@ in
       subdomain = "auth";
       ssl = config.shb.certs.certs.selfsigned.n;
 
-      ldapEndpoint = "ldap://127.0.0.1:${builtins.toString config.shb.ldap.ldapPort}";
+      ldapHostname = "127.0.0.1";
+      ldapPort = config.shb.ldap.ldapPort;
       dcdomain = config.shb.ldap.dcdomain;
 
       secrets = {